refactor : 네이버 클라우드 코드 변경
